requirements.txt
email-validator==2.0.0.post2
fastapi==0.95.2
motor==3.1.2
pydantic==1.10.7
python-dotenv==1.0.0
uvicorn==0.22.0